+++ HISTORY.rst
@@ -2,10 +2,11 @@
 History
 =======
 
-Unreleased
-----------
+1.0.0 (2024-08-08)
+------------------
 
-* Add progress bar to pages.
+* Fix button customization in paged surveys.
+* Add progress bar option to pages.
 * Add option to pass a list to Page's constructor, in order to both have the current page number and an associated value.
 * Add survey download button and file importer.
 * Add "widget_key" to survey data to help restore survey data for displayed widgets.